addressing comments PR
